% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_means.R
\name{get_M_samples}
\alias{get_M_samples}
\title{Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.}
\usage{
get_M_samples(
  pop.M = NA,
  pop.SD = NA,
  n = 10,
  number.of.samples = 10,
  number.of.decimals = 2,
  expanded.output = FALSE,
  seed.value = 1,
  n.min = NULL,
  n.max = NULL,
  label = NULL
)
}
\arguments{
\item{pop.M}{Population mean}

\item{pop.SD}{Population SD}

\item{n}{Cell size for both cells for all samples. If you use n, do not use n.min or n.max.}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}

\item{seed.value}{random number seed value}

\item{n.min}{Use various sample sizes. Uniform distribution min.}

\item{n.max}{Use various sample sizes. Uniform distribution max}

\item{label}{label for all studies}

\item{pop.data}{Data to use for population (optional)}

\item{pop.column.name}{Column name from data to use for sampling (optional)}
}
\value{
Data frame with desired properties
}
\description{
Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.
}
\examples{
get_M_samples(pop.M = 100, pop.SD = 15, n = 100)
}
