% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_middle_percent}
\alias{get_middle_percent}
\title{Calculate the percentage of rows that are true in a column}
\usage{
get_middle_percent(sample_statistic, middle_percent = 95)
}
\arguments{
\item{sample_statistic}{the column to be examined}

\item{middle_percent}{the desired middle percentage (default is 95)}
}
\value{
values that bound the specified middle percent
}
\description{
Calculate the percentage of rows that are true in a column
}
