% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_correlations.R
\name{get_r_samples}
\alias{get_r_samples}
\title{Calculate a number of sample correlations based on a specified population correlation}
\usage{
get_r_samples(
  pop.r = NA,
  n,
  number.of.samples = 100,
  number.of.decimals = 3,
  seed.value = NULL
)
}
\arguments{
\item{pop.r}{Population correlation. Do not use pop.data is you provide this value.}

\item{n}{Sample size for all samples If you use n, do not use n.min or n.max.}

\item{number.of.samples}{Number of samples to obtain}

\item{number.of.decimals}{Number of decimals to report in returned data frame}

\item{seed.value}{random number seed value}
}
\value{
Data frame with sample correlations
}
\description{
Calculate a number of sample correlations based on a specified population correlation
}
\examples{
get_r_samples(pop.r = .35,n=100)
}
