% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_functions.R
\name{nonpivotal_ci_d_indep}
\alias{nonpivotal_ci_d_indep}
\title{Create animation of nonpivotal CI for d-independent groups}
\usage{
nonpivotal_ci_d_indep(
  d,
  n.1,
  n.2,
  filename = "video-d-indep.mp4",
  level = 0.95
)
}
\arguments{
\item{d}{sample d value}

\item{n.1}{sample size}

\item{n.2}{sample size}

\item{filename}{filename for video (default: video-d-indep.mp4). Must end in .mp4}

\item{level}{confidence level (default: .95)}
}
\value{
ggplot object
}
\description{
Create animation of nonpivotal CI for d-independent groups
}
