% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_functions.R
\name{nonpivotal_ci_d_rep}
\alias{nonpivotal_ci_d_rep}
\title{Create animation of nonpivotal CI for d-repeated measures}
\usage{
nonpivotal_ci_d_rep(d, n, filename = "output.mp4", level = 0.95)
}
\arguments{
\item{d}{sample d value}

\item{n}{sample size}

\item{filename}{filename for video (default: output.mp4). Must end in .mp4}

\item{level}{confidence level (default: .95)}
}
\value{
ggplot object
}
\description{
Create animation of nonpivotal CI for d-repeated measures
}
