% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_functions.R
\name{nonpivotal_ci_r}
\alias{nonpivotal_ci_r}
\title{Create animation of nonpivotal CI for correlation}
\usage{
nonpivotal_ci_r(r, n, filename = "video-r.mp4", level = 0.95)
}
\arguments{
\item{r}{sample d value}

\item{n}{sample size}

\item{filename}{filename for video (default: video-d-indep.mp4). Must end in .mp4}

\item{level}{confidence level (default: .95)}
}
\value{
ggplot object
}
\description{
Create animation of nonpivotal CI for correlation
}
