% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{p_range_count}
\alias{p_range_count}
\title{Probability of a p-value in a specified range}
\usage{
p_range_count(data, min, max)
}
\arguments{
\item{data}{upper end of range (less than)}

\item{min}{lower end of range (equal to or greater than)}
}
\value{
The probability
}
\description{
Probability of a p-value in a specified range
}
