% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{percent_true}
\alias{percent_true}
\title{Calculate the percentage of rows that are true in a column}
\usage{
percent_true(x)
}
\arguments{
\item{x}{the column to be examined}
}
\value{
The percent equal to TRUE
}
\description{
Calculate the percentage of rows that are true in a column
}
