% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plot_ci}
\alias{plot_ci}
\title{Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.}
\usage{
plot_ci(
  data,
  capture.colors = c("red", "black"),
  pop.line.color = "blue",
  pop.line.width = 1.5,
  ...
)
}
\arguments{
\item{data}{simulation data}

\item{capture.colors}{colors to shade intervals that capture or do not capture parameter}

\item{pop.line.color}{color of vertical line for parameter}

\item{pop.line.width}{thickness of vertical line for parameter}
}
\value{
ggplot object
}
\description{
Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.
}
