% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.}
\usage{
plot_distribution(.data, type = "text", ...)
}
\arguments{
\item{.data}{simulation data}

\item{type}{Use "text" for histogram with letters or "bar" for standard histogram}
}
\value{
ggplot object
}
\description{
Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.
}
