% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plot_pvalues}
\alias{plot_pvalues}
\title{Plot p-values for a single distribution}
\usage{
plot_pvalues(
  .data,
  type.1.error.prob = 0.05,
  sig.line.color = "green",
  sig.line.width = 1,
  ...
)
}
\arguments{
\item{.data}{simulation data}

\item{type.1.error.prob}{your choice of alpha. default is .05}

\item{sig.line.color}{color of vertical line for significance}

\item{sig.line.width}{thickness of vertical line for significance}
}
\value{
ggplot object
}
\description{
Plot p-values for a single distribution
}
