% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plot_compare_pvalues}
\alias{plot_compare_pvalues}
\title{Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.}
\usage{
plot_compare_pvalues(
  null_dist,
  eff_dist,
  type.1.error.prob = 0.05,
  sig.line.color = "green",
  sig.line.width = 1,
  ...
)
}
\arguments{
\item{type.1.error.prob}{your choice of alpha. default is .05}

\item{sig.line.color}{color of vertical line for significance}

\item{sig.line.width}{thickness of vertical line for significance}

\item{data}{simulation data}
}
\value{
ggplot object
}
\description{
Calculate a number of sample d-values (unbiased) based on a specified (infinite) population correlation.
}
